Move quat, rect2, transform2d -> core_types::geom (internal-only)
