+++ gdnative-core/src/core_types/geom/quat.rs
