+++ gdnative-core/src/core_types/mod.rs
@@ -17,12 +17,9 @@ mod float32_array;
 mod int32_array;
 mod node_path;
 mod pool_array;
-mod quat;
-mod rect2;
 mod rid;
 mod string;
 mod string_array;
-mod transform2d;
 mod variant;
 mod variant_array;
 mod vector2;
